+++ CHANGELOG.md
@@ -5,6 +5,9 @@
   necessarily depending on the newest version of the compiler. In addition
   to that Ormolu is now GHCJS-compatible.
 
+* Now unrecognized GHC options passed with `--ghc-opt` cause Ormolu to fail
+  (exit code 7).
+
 * Fixed formatting of result type in closed type families. See [issue
   420](https://github.com/tweag/ormolu/issues/420).